// src/enums/dashboardEnum.ts
export enum DashboardRpmEnum {
	/* 最小转速，单位 {1/min * 1000} */
	minRpm = 0.8,
	/* 最大转速，单位 {1/min * 1000} */
	maxRpm = 8
}

export enum DashboardSpeedEnum {
	/* 最小迈速，单位 {km/h} */
	minSpeed = 0,
	/* 最大迈速，单位 {km/h} */
	maxSpeed = 300
}

export enum DashboardGearEnum {
	/* 档位 1 */
	gearOne = 'D1',
	/* 档位 2 */
	gearTwo = 'D2',
	/* 档位 3 */
	gearThree = 'D3',
	/* 档位 4 */
	gearFour = 'D4',
	/* 档位 5 */
	gearFive = 'D5',
	/* 档位 6 */
	gearSix = 'D6',
	/* 档位 7 */
	gearSeven = 'D7',
	/* 档位 8 */
	gearEight = 'D8'
}